Use VictoriaMetrics as time series database (#5)
InfluxDB has created with version 3 a lot of disturbances. They will
drastically alter their business model and announced a community
version, which is not seen yet. Thus many stick to version 1 or 2 of
InfluxDB.

Lets instead switch to VictoriaMetrics which is compatible, available
and maintained. On top of that it claims to have better performance and
to use less resources.